requirements.txt
Django==4.1.1
djangorestframework==3.13.1
Pillow==9.2.0
requests==2.28.1
django-cors-headers==3.13.0
mysqlclient==2.1.1
PyMySQL==1.0.2
ipython==8.5.0